Add Makefile, configuration files and scripts
Update the README with instructions on setting up, configuring, building
and running the Redis app with Unikraft. As part of it, add Makefile
files, configuration files and scripts; they are referenced in the
`README.md` file.

Current instructions cover the use of QEMU/KVM platform (both on x86_64 and
on AArch64) and GCC. Firecraker/KVM, Xen, Linux platforms and Clang are
not yet documented.

Signed-off-by: Ioan-Teodor Teugea <ioan_teodor.teugea@stud.acs.upb.ro>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #16