Add Firecracker support
Add configuration files and support script to configure, build and run
Redis with Unikraft on Firecracker. This can be built for x86_64 and
initrd.

Note that, as of this commit (for release 0.14), networking is not yet
integrated in Firecracker. So a Redis server will start as a Unikraft
Firecracker instance, but a network connection won't be (yet) possible.

Signed-off-by: Razvan Deaconescu <razvand@unikraft.io>
Reviewed-by: Radu Nichita <radunichita99@gmail.com>
Approved-by: Stefan Jumarea <stefanjumarea02@gmail.com>
GitHub-Closes: #18